Add New Visual Effects
-New godray effect via particle system. It can be easily adjust to
improve its performance.
-New film play object. Now film plane can be attached to the object. So
in the AR scene, film plane will be moved base on the movement of
recognized object. Also, the film plane doesn't have to occupy the
entire screen. The size of it can be easily adjusted.
-Film transition animation. You can see it in scene "godray 1". Just
press play and with. Its performance needs to be improved, but it's
pretty cool now.